Initial UI and full/partial refresh

+++ src/main/kotlin/com/runt9/cashflow/service/BankService.kt
@@ -1,13 +1,17 @@
 package com.runt9.cashflow.service
 
+import com.runt9.cashflow.model.RefreshType
 import com.runt9.cashflow.model.dto.BankLogin
+import com.runt9.cashflow.model.entity.Account
 import com.runt9.cashflow.model.entity.Bank
+import com.runt9.cashflow.model.entity.Transaction
 import com.runt9.cashflow.repository.AccountRepository
 import com.runt9.cashflow.repository.BankRepository
 import com.runt9.cashflow.repository.TransactionRepository
 import com.runt9.cashflow.service.dataGatherer.DataGathererFactory
 import org.apache.tomcat.util.codec.binary.Base64
 import org.springframework.stereotype.Service
+import java.time.LocalDate
 import javax.crypto.Cipher
 import javax.crypto.spec.SecretKeySpec
 
@@ -44,11 +48,11 @@ class BankService(
         return BankLogin(bank.bankType, loginPair.substring(0, loginNameLen), loginPair.substring(loginNameLen))
     }
 
-    // TODO: Full and partial refresh
-    fun refreshBanks(encryptionKey: String) {
-        // TODO: Remove these and properly update
-        accountRepository.deleteAll()
-        transactionRepository.deleteAll()
+    fun refreshBanks(encryptionKey: String, refreshType: RefreshType) {
+        if (refreshType == RefreshType.FULL) {
+            accountRepository.deleteAll()
+            transactionRepository.deleteAll()
+        }
 
         val banks = bankRepository.findAll()
         banks.forEach {
@@ -57,12 +61,52 @@ class BankService(
             try {
                 dataGatherer.login(login.loginName, login.password)
                 val accounts = dataGatherer.gatherAccounts(it)
-                // TODO: Create or update
+
+                if (refreshType == RefreshType.FULL) {
+                    accountRepository.save(accounts)
+                    transactionRepository.save(accounts.flatMap { dataGatherer.getAccountTransactions(it, refreshType) })
+                } else {
+                    accountRepository.save(mergeAccounts(accounts))
+                    transactionRepository.save(accounts.flatMap { mergeTransactions(it, dataGatherer.getAccountTransactions(it, refreshType)) })
+                }
+
+                accounts.forEach { it.lastRefresh = LocalDate.now() }
                 accountRepository.save(accounts)
-                transactionRepository.save(accounts.flatMap { dataGatherer.getAccountTransactions(it) })
             } finally {
                 dataGatherer.cleanup()
             }
         }
+    }
+
+    private fun mergeTransactions(account: Account, transactions: List<Transaction>): List<Transaction> {
+        val existingTransactions = transactionRepository.findByAccount(account)
+        val mergedTransactions = arrayListOf<Transaction>()
+
+        transactions.forEach { newTran ->
+            val existingTran = existingTransactions.find { it.merchant.id == newTran.merchant.id && it.date == newTran.date && it.amount == newTran.amount }
+            if (existingTran != null) {
+                newTran.id = existingTran.id
+            }
+
+            mergedTransactions += newTran
+        }
+
+        return mergedTransactions
+    }
+
+    fun mergeAccounts(accounts: List<Account>): List<Account> {
+        val existingAccounts = accountRepository.findAll()
+        val mergedAccounts = arrayListOf<Account>()
+        
+        accounts.forEach { newAcc ->
+            val existingAcc = existingAccounts.find { it.name == newAcc.name }
+            if (existingAcc != null) {
+                newAcc.id = existingAcc.id
+            }
+
+            mergedAccounts += newAcc
+        }
+        
+        return mergedAccounts
     }
 }
